Added ramp up to bang bang control.
-The ramp up triggers using this if statement
if (error < -tp100ms / 2)


tp100ms is the target ticks per 100 milliseconds.
error is the deviation from this target.
(If the tp100ms is 300, and the actual speed the shooter is turning at is 50, error is -250.

if the tp100ms is 300, and the actual speed of the shooter is 400 however, the error would be 100.)

As a result, it works like this:
If our error is -300 and our tp100ms is 300, our
-tp100ms is -300.
divide that by 2 and you get -150.
-300 is less than -150, so the shooter will ramp up.

However if our error is 300, 300 is definitely MORE than -150,
so it will continue on to the other if statements.